+++ packages/runtime/src/util/sourcemap-errors.ts
@@ -26,8 +26,8 @@ export default async (job: Job, error: RTError) => {
     }
 
     if (error.pos && !isNaN(error.pos.line!)) {
-      // TODO how to handle file name properly here?
-      const src = smc.sourceContentFor('src.js')!.split('\n');
+      const fileName = job.name ? `${job.name}.js` : 'src.js';
+      const src = smc.sourceContentFor(fileName)!.split('\n');
       const line = src[error.pos.line! - 1];
       error.pos.src = line;
     }